+++ packages/flutter_markdown/lib/src/builder.dart
@@ -2,14 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// ignore: unnecessary_import, see https://github.com/flutter/flutter/pull/138881
-import 'dart:ui';
-
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:markdown/markdown.dart' as md;
 
-import '_functions_io.dart' if (dart.library.html) '_functions_web.dart';
+import '_functions_io.dart' if (dart.library.js_interop) '_functions_web.dart';
 import 'style_sheet.dart';
 import 'widget.dart';
 
@@ -867,8 +864,7 @@ class MarkdownBuilder implements md.NodeVisitor {
     if (selectable) {
       return SelectableText.rich(
         text!,
-        // ignore: deprecated_member_use
-        textScaleFactor: styleSheet.textScaleFactor,
+        textScaler: styleSheet.textScaler,
         textAlign: textAlign ?? TextAlign.start,
         onTap: onTapText,
         key: k,
@@ -876,8 +872,7 @@ class MarkdownBuilder implements md.NodeVisitor {
     } else {
       return Text.rich(
         text!,
-        // ignore: deprecated_member_use
-        textScaleFactor: styleSheet.textScaleFactor,
+        textScaler: styleSheet.textScaler,
         textAlign: textAlign ?? TextAlign.start,
         key: k,
       );