+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.12
+
+* Converts `refreshReceipt()`, `startObservingPaymentQueue()`, `stopObservingPaymentQueue()`, 
+`registerPaymentQueueDelegate()`, `removePaymentQueueDelegate()`, `showPriceConsentIfNeeded()` to pigeon.
+
+## 0.3.11
+
+* Fixes SKError.userInfo not being nullable.
+
 ## 0.3.10
 
 * Converts `startProductRequest()`, `finishTransaction()`, `restoreTransactions()`, `presentCodeRedemptionSheet()` to pigeon.